Move components to folders. Auth route
