+++ src/views/Auth.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function index() {
+  return (
+    <div>Auth</div>
+  )
+}
